dz_1_task_1
