+++ index.html
@@ -30,7 +30,7 @@
 
 
 
-<section id="timerSection" class="visible">
+<section id="timerSection" class="invisible">
   <h3>Таймер</h3>
 <!--  <form id="timerForm">-->
     <label>
